% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_names.R
\name{all_display_names}
\alias{all_display_names}
\title{Return all display names for endogenous variables}
\usage{
all_display_names(M_)
}
\arguments{
\item{M_}{Dynare model object containing `endo.names`.}
}
\value{
Named character vector mapping display names to Dynare names.
}
\description{
Return all display names for endogenous variables
}
