% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_alt_paths_tab.R
\name{alt_paths_baseline_template}
\alias{alt_paths_baseline_template}
\title{Build an upload template from a canonical baseline.}
\usage{
alt_paths_baseline_template(file, baseline)
}
\arguments{
\item{file}{Destination `.xlsx` file.}

\item{baseline}{Canonical `ezdyn_baseline` object.}
}
\value{
Invisibly, `NULL`.
}
\description{
Build an upload template from a canonical baseline.
}
