% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_alt_paths_tab.R
\name{alt_paths_default_plot_range}
\alias{alt_paths_default_plot_range}
\title{Return the initial Alternative Paths plot range.}
\usage{
alt_paths_default_plot_range(timeline)
}
\arguments{
\item{timeline}{Validated dashboard timeline.}
}
\value{
Two `Date` values for the plot start and end.
}
\description{
Return the initial Alternative Paths plot range.
}
