% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_alt_paths_tab.R
\name{alt_paths_plot_title}
\alias{alt_paths_plot_title}
\title{Return an Alternative Paths graph title from the selected responses.}
\usage{
alt_paths_plot_title(display_names)
}
\arguments{
\item{display_names}{Display names present in the plot data.}
}
\value{
One plot title.
}
\description{
Return an Alternative Paths graph title from the selected responses.
}
