% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_cash_rate_slider.R
\name{alt_paths_slider_label}
\alias{alt_paths_slider_label}
\title{Format one forecast date as a year-quarter slider label.}
\usage{
alt_paths_slider_label(date)
}
\arguments{
\item{date}{One quarterly `Date` value.}
}
\value{
One label such as `"26-Q2"`.
}
\description{
Format one forecast date as a year-quarter slider label.
}
