% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_alt_paths_tab.R
\name{alt_paths_upload_ui}
\alias{alt_paths_upload_ui}
\title{Build the configured-baseline upload trigger.}
\usage{
alt_paths_upload_ui(ns, button_label = "Provide Custom Baseline")
}
\arguments{
\item{ns}{A Shiny namespace function.}

\item{button_label}{Upload trigger label.}
}
\value{
Shiny UI.
}
\description{
Build the configured-baseline upload trigger.
}
