% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_irf.R
\name{calc_scale_factors}
\alias{calc_scale_factors}
\title{Calculate scaling factors for all variables in metadata.}
\usage{
calc_scale_factors(expr_df, M_)
}
\arguments{
\item{expr_df}{Data frame with a `scale_formula` column.}

\item{M_}{Model structure containing `param_df`.}
}
\value{
Metadata data frame with a new `scale_factor` column.
}
\description{
Calculate scaling factors for all variables in metadata.
}
