% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_cash_rate_slider.R
\name{cash_rate_slider_styles}
\alias{cash_rate_slider_styles}
\title{Return styles for the dashboard and cash-rate slider component.}
\usage{
cash_rate_slider_styles()
}
\value{
HTML head tags containing dashboard CSS.
}
\description{
Return styles for the dashboard and cash-rate slider component.
}
