% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_names.R
\name{defined_display_names}
\alias{defined_display_names}
\title{Get only those display names which are defined in the variable metadata.
TEMPORARY 'quick and dirty' function for until this implementation is reworked.}
\usage{
defined_display_names(M_, defined_only = T, alt_paths_only = F, hsd_only = F)
}
\arguments{
\item{alt_paths_only}{whether to only pull variables with alt paths == T in the variable metadata.}

\item{hsd_only}{whether to only pull variables with hsd == T in the variable metadata.}
}
\description{
Get only those display names which are defined in the variable metadata.
TEMPORARY 'quick and dirty' function for until this implementation is reworked.
}
