% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_names.R
\name{display_name}
\alias{display_name}
\title{Get the display name for a Dynare variable}
\usage{
display_name(dynare_name_, M_)
}
\arguments{
\item{dynare_name_}{Character scalar. Dynare variable name.}

\item{M_}{Dynare model object containing `varmeta`.}
}
\value{
Character scalar display name.
}
\description{
Returns the display name for a single Dynare variable. If multiple
transformations of a variable are defined in the metadata, this returns the default.
}
\examples{
\dontrun{
display_name("infl_obs", M_)
}
}
