% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_names.R
\name{display_names}
\alias{display_names}
\title{Get display names for multiple Dynare variables}
\usage{
display_names(dynare_names, M_, df = F)
}
\arguments{
\item{dynare_names}{Character vector of Dynare variable names.}

\item{M_}{Dynare model object containing `varmeta`.}

\item{df}{Logical. If TRUE, return a data frame with columns `Variable` and
`display_name`.}
}
\value{
Named character vector or tibble.
}
\description{
Get display names for multiple Dynare variables
}
\examples{
\dontrun{
display_names(c("infl_obs", "r_obs"), M_)
}
}
