% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_names.R
\name{display_names_and_units}
\alias{display_names_and_units}
\title{Get display names, units, and unit symbols}
\usage{
display_names_and_units(
  dynare_names,
  M_,
  unit_symbol_irf = F,
  unit_symbol_baseline = F
)
}
\arguments{
\item{dynare_names}{Character vector of Dynare variable names.}

\item{M_}{Dynare model object containing `varmeta`.}

\item{unit_symbol_irf}{Whether to include the unit symbols (IRF) in the dataframe.}

\item{unit_symbol_baseline}{Whether to include the unit symbol (baseline) in the dataframe.}
}
\value{
Tibble with columns for display name, unit, and unit symbols.
}
\description{
Get display names, units, and unit symbols
}
