% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hd.R
\name{extract_dynare_names}
\alias{extract_dynare_names}
\title{Convert Dynare-style name containers to a clean character vector.
TODO: deprecate.}
\usage{
extract_dynare_names(x)
}
\description{
Convert Dynare-style name containers to a clean character vector.
TODO: deprecate.
}
