% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_hd.R
\name{ez_hd.available_shock_groupings}
\alias{ez_hd.available_shock_groupings}
\title{Returns what shock groupings are available for use with ez_hd.}
\usage{
ez_hd.available_shock_groupings(M_)
}
\arguments{
\item{M_}{dynare M_ object.}
}
\description{
Returns what shock groupings are available for use with ez_hd.
}
