% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_hd.R
\name{ez_hd}
\alias{ez_hd}
\title{Convert historical decomposition output to a tidy table.}
\usage{
ez_hd(M_, oo_ = NULL, hd_override = NULL, model_name = NULL)
}
\arguments{
\item{M_}{Dynare model object used to retrieve display names/units metadata
analogously to `pretty_irf()`.}

\item{oo_}{oo_ object. If hd_override=NULL, the shock decomposition will be generated
by calling get_hd on M_ and oo_. This input will not be used if hd_override is not
NULL.}

\item{hd_override}{Supply HD manually based on either output of `get_hd()` or even
run_hd.m MATLAB output. If hd_override is supplied, oo_ can be NULL. Either:
  - a list containing `shock_decomposition` (e.g. output of `get_hd()`), or
  - a 3D array with dimensions `[variable x shock x time]`.
  - a long dataframe containing columns (`t`, `shock`, `variable`, `value`).
     This could have been outputted from the MATLAB version of run_hd.}

\item{model_name}{Optional model name override; defaults to `M_$model_name`
when available.}
}
\value{
A long data frame with at least columns:
  `t`, `shock`, `variable`, `model_name`, `display_name`, `display_unit`, `value`.
  Additional shock metadata columns are added from `M_$shock_meta` if available.
}
\description{
Convert historical decomposition output to a tidy table.
}
\details{
TODO: Need to apply rescaling to the shock output.
}
