% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_tables.R
\name{ezdyn_attach_param_table}
\alias{ezdyn_attach_param_table}
\title{Build enriched parameter table for model object}
\usage{
ezdyn_attach_param_table(M_, path_meta = NULL)
}
\arguments{
\item{M_}{Model object.}

\item{path_meta}{Optional path to metadata workbook.}
}
\value{
Tibble containing enriched parameter metadata and values.
}
\description{
Build enriched parameter table for model object
}
