% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_tables.R
\name{ezdyn_param_values_from_M}
\alias{ezdyn_param_values_from_M}
\title{Build parameter values table from model object}
\usage{
ezdyn_param_values_from_M(M_)
}
\arguments{
\item{M_}{Model object (Dynare or custom moo).}
}
\value{
Tibble with parameter values (`Name`, `Value`, `TexName`).
}
\description{
Build parameter values table from model object
}
