% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_ir_matrix.R
\name{flatten_ir_matrix}
\alias{flatten_ir_matrix}
\title{Given an IR matrix from get_ir_matrix(), flatten the matrix so that each row is a response variable for a particular time period, and each 
column is a shock that occurred in a particular time period. 
This is useful for backing out the scale of shocks required to 
achieve a given impulse response.}
\usage{
flatten_ir_matrix(
  ir_matrix,
  selected_resp_vars = dimnames(ir_matrix)[[1]],
  selected_periods = 1:dim(ir_matrix)[[2]]
)
}
\arguments{
\item{ir_matrix}{result from get_ir_matrix()}

\item{selected_resp_vars}{if you want to subset the variables before flattening. Default is selects all response variables.}

\item{selected_periods}{Subset the number of periods before flattening. Default selects all periods in ir_matrix.}
}
\description{
Given an IR matrix from get_ir_matrix(), flatten the matrix so that each row is a response variable for a particular time period, and each 
column is a shock that occurred in a particular time period. 
This is useful for backing out the scale of shocks required to 
achieve a given impulse response.
}
\examples{
ir_matrix <- get_ir_matrix(M_, oo_, 40, list("eps_r"=1:3, "eps_psi"=1:3))
flat_matrix <- ir_matrix[rownames(ir_matrix) \%in\% c("r_obs", "ntwi_growth"),,,] |>
    flatten_ir_matrix() 

}
