% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_ir_matrix.R
\name{flatten_target}
\alias{flatten_target}
\title{Given a target matrix, flatten it so that each row is a response variable for a particular time period.}
\usage{
flatten_target(target)
}
\description{
Given a target matrix, flatten it so that each row is a response variable for a particular time period.
}
\examples{
target <- matrix(0, nrow=2, ncol=40) 
rownames(target) = c("r_obs", "ntwi_growth")
target
}
