% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dr_matrices.R
\name{get_dr_matrices}
\alias{get_dr_matrices}
\title{Get decision rule coefficients for lagged variables and structural shocks.}
\usage{
get_dr_matrices(M_, oo_, adjusted = TRUE)
}
\arguments{
\item{M_}{Dynare model structure containing model specifications and parameters.}

\item{oo_}{Dynare output structure containing simulation results.}

\item{adjusted}{whether to adjust or otherwise return the raw matrices from dynare. 
Adjustment implies reordering A and B, as well as restricting A to only contain state variables.}
}
\description{
Get decision rule coefficients for lagged variables and structural shocks.
}
