% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IR_matrix.R
\name{get_ir_matrix}
\alias{get_ir_matrix}
\title{Create a MxHxSxT matrix where:
M: Number of response variables to IRF 
H: Number of time periods (i.e. horizon)
S: Number of shock variables.
T: Maximum period in which a shock occurs.}
\usage{
get_ir_matrix(M_, oo_, horizon = 40, shock_timing)
}
\arguments{
\item{horizon}{horizon for which the irf is plotted.}

\item{shock_timing}{Named list. Name is the name of the shock, values are a list of periods in which the shock is active.}
}
\description{
Create a MxHxSxT matrix where:
M: Number of response variables to IRF 
H: Number of time periods (i.e. horizon)
S: Number of shock variables.
T: Maximum period in which a shock occurs.
}
