% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ir_matrix_gabaix.R
\name{get_ir_matrix_gabaix}
\alias{get_ir_matrix_gabaix}
\title{Get IRF matrix for partially anticipated shock, using cognitive discounting.
This is inspired by the method described in Gabaix 2020.
The discounting is applied to the shock (not the state vector as in Gabaix 2020).
In depth: 
We have a partially anticipated shock which is announced in period 1. 
This is equivalent to some linear combination of fully anticipated 
shocks announced in periods 1 to T_shock (i.e. s_ann=1:T_shock) and 
with anticipation horizon h_ant=1 to T_shock.}
\usage{
get_ir_matrix_gabaix(M_, oo_, horizon, shock_name, T_shock, lambda)
}
\arguments{
\item{M_}{M_ object.}

\item{oo_}{oo_ object.}

\item{horizon}{number of periods to compute IRF for.}

\item{shock_name}{name of the shock}

\item{T_shock}{}

\item{lambda}{Gabaix discounting parameter. This is the fraction of people who become aware of the shock in each period.}
}
\description{
Get IRF matrix for partially anticipated shock, using cognitive discounting.
This is inspired by the method described in Gabaix 2020.
The discounting is applied to the shock (not the state vector as in Gabaix 2020).
In depth: 
We have a partially anticipated shock which is announced in period 1. 
This is equivalent to some linear combination of fully anticipated 
shocks announced in periods 1 to T_shock (i.e. s_ann=1:T_shock) and 
with anticipation horizon h_ant=1 to T_shock.
}
\examples{
ir_matrix_discounted <- get_ir_matrix_gabaix(M_, oo_, 40, "eps_r", 5, 0.8)
ir_matrix_discounted$Mh_total[1,1,1,1]

}
