% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_irf.R
\name{get_irf}
\alias{get_irf}
\title{Get the impulse response function for given shocks. 
If array is true, returns the following MxHxS array: 
M: Number of response variables to IRF 
H: Number of time periods (i.e. horizon)
S: Number of shock variables.}
\usage{
get_irf(M_, oo_, shock_names, horizon, pretty = FALSE, data_frame = pretty)
}
\arguments{
\item{M_}{Model structure containing model metadata.}

\item{oo_}{Model output object.}

\item{shock_names}{Character vector of shock codes. If `M_$shock_meta` has a
`description` column, descriptions can also be supplied.}

\item{horizon}{Horizon to calculate impulse responses over.}

\item{pretty}{Boolean. If TRUE, returns the rescaled IRFs with names and units in a nice dataframe. 
Rescaling, names and units are based on metadata in M_.}

\item{data_frame}{Logical. If TRUE, return a data frame instead of an array.}
}
\description{
Get the impulse response function for given shocks. 
If array is true, returns the following MxHxS array: 
M: Number of response variables to IRF 
H: Number of time periods (i.e. horizon)
S: Number of shock variables.
}
\details{
`M_` and `oo_` must have matching `dynare` or `custom_moo` classes.
Load non-Dynare models as a custom MOO and pass its `M_` and `oo_` objects explicitly.
}
