% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_irf.R
\name{get_irf_dynare}
\alias{get_irf_dynare}
\title{Compute IRFs from Dynare decision-rule matrices.}
\usage{
get_irf_dynare(M_, oo_, shock_names, horizon, data_frame = FALSE)
}
\arguments{
\item{M_}{Dynare model structure.}

\item{oo_}{Dynare model output structure.}

\item{shock_names}{Character vector of Dynare shock codes.}

\item{horizon}{Horizon to calculate impulse responses over.}

\item{data_frame}{Logical. If TRUE, return a data frame instead of an array.}
}
\description{
Compute IRFs from Dynare decision-rule matrices.
}
\keyword{internal}
