% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_names.R
\name{get_shiny_displaynames}
\alias{get_shiny_displaynames}
\title{Get display-name lists for the Shiny selector}
\usage{
get_shiny_displaynames(
  M_,
  core_only = F,
  alt_paths_only = F,
  extra_names = list()
)
}
\arguments{
\item{M_}{Dynare model object containing `varmeta` and `endo.names`.}

\item{core_only}{show only the core variables}

\item{alt_paths_only}{Logical. If TRUE, only return variables with
`alt_paths == TRUE`.}

\item{extra_names}{Any extra names to force-add to the list of displaynames.
This is used in the dashboard to force add the Unemployment rate as an extra
variable (since it isn't in DINGO).}

\item{alt_paths}{only return variables with alt_paths=TRUE in the variable metadata}
}
\value{
Named list of character vectors.
}
\description{
Returns a list of named vectors for Shiny inputs, split by core and other
variables and optionally filtered to alternative-path variables.
}
