% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline.R
\name{getBaseline}
\alias{getBaseline}
\title{Generate baseline profile. Some of this code is not relevant since we aren't
extending the forecast.}
\usage{
getBaseline(M_, oo_, data_start, data_end)
}
\arguments{
\item{M_}{MATLAB M_ object.}

\item{oo_}{oo_ object.}

\item{data_start}{Start date of the data sample inside oo_.}

\item{data_end}{End date of the data sample inside oo_.}
}
\description{
Generate baseline profile. Some of this code is not relevant since we aren't
extending the forecast.
}
