% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_hd.R
\name{hd_matrix_to_df}
\alias{hd_matrix_to_df}
\title{Convert HD array to long data frame.}
\usage{
hd_matrix_to_df(arr)
}
\arguments{
\item{arr}{3D array with dimensions [variable x shock x time].}
}
\value{
Data frame with columns `variable`, `shock`, `t`, `value`.
}
\description{
Convert HD array to long data frame.
}
