% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_help_widgets.R
\name{help_button}
\alias{help_button}
\title{Build a small circular help-modal trigger button.}
\usage{
help_button(id, circle = TRUE)
}
\arguments{
\item{id}{Input identifier.}

\item{circle}{Whether to use a compact circular button style.}
}
\value{
Shiny UI.
}
\description{
Build a small circular help-modal trigger button.
}
