% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_hsd_tab.R
\name{hsd_choices}
\alias{hsd_choices}
\title{Return valid named Shiny choices.}
\usage{
hsd_choices(values)
}
\arguments{
\item{values}{Character values.}
}
\value{
A named character vector.
}
\description{
Return valid named Shiny choices.
}
