% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_help_widgets.R
\name{infoButton}
\alias{infoButton}
\title{Build a popover info button.}
\usage{
infoButton(id, title, msg)
}
\arguments{
\item{id}{Input identifier.}

\item{title}{Popover title.}

\item{msg}{Popover message.}
}
\value{
Shiny UI.
}
\description{
Build a popover info button.
}
