% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_help_widgets.R
\name{infoTooltip}
\alias{infoTooltip}
\title{Build a tooltip info icon.}
\usage{
infoTooltip(id, title, msg)
}
\arguments{
\item{id}{Unused; retained for call-site symmetry with [infoButton()].}

\item{title}{Unused; retained for call-site symmetry with [infoButton()].}

\item{msg}{Tooltip message.}
}
\value{
Shiny UI.
}
\description{
Build a tooltip info icon.
}
