% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plot.R
\name{irf_build_unit_symbols}
\alias{irf_build_unit_symbols}
\title{Build a named list of plot unit symbols for facet panels.}
\usage{
irf_build_unit_symbols(df)
}
\arguments{
\item{df}{Data frame created by [pretty_irf()] or [get_alt_paths()].}
}
\value{
Named list mapping `display_name` to unit symbols.
}
\description{
Build a named list of plot unit symbols for facet panels.
}
