% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_irf_helpers.R
\name{irf_choices}
\alias{irf_choices}
\title{Return valid named Shiny choices.}
\usage{
irf_choices(values)
}
\arguments{
\item{values}{Character values.}
}
\value{
A named character vector.
}
\description{
Return valid named Shiny choices.
}
