% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plot.R
\name{irf_filter_pretty}
\alias{irf_filter_pretty}
\title{Filter pretty IRF data by response names and shocks.}
\usage{
irf_filter_pretty(df, dynare_names = NULL, display_names = NULL, shocks = NULL)
}
\arguments{
\item{df}{Data frame created by `pretty_irf`.}

\item{dynare_names}{Optional vector of dynare names to include.}

\item{display_names}{Optional vector of display names to include.}

\item{shocks}{Optional vector of shocks to include.}
}
\value{
Filtered data frame.
}
\description{
Filter pretty IRF data by response names and shocks.
}
