% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_optimal_policy_tab.R
\name{optimal_policy_response_choices}
\alias{optimal_policy_response_choices}
\title{Return available policy response choices from model metadata.}
\usage{
optimal_policy_response_choices(model)
}
\arguments{
\item{model}{Full policy-model MOO pair.}
}
\value{
Named character vector from display names to model codes.
}
\description{
Return available policy response choices from model metadata.
}
