% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_tables.R
\name{param_table_katex}
\alias{param_table_katex}
\title{Render KaTeX parameter table from model object}
\usage{
param_table_katex(M_, type = NULL, dp = 3, as_of = NULL)
}
\arguments{
\item{M_}{Model object containing `param_table`.}

\item{type}{Optional value to filter by `Type` column.}

\item{dp}{Number of decimal places for rounding numeric values.}

\item{as_of}{Optional timestamp text appended as a table footnote.}
}
\value{
HTML `kableExtra` table.
}
\description{
Render KaTeX parameter table from model object
}
