% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plot.R
\name{pretty_irf}
\alias{pretty_irf}
\title{Convert IRF into a pretty format with names and units.}
\usage{
pretty_irf(irf_data, M_)
}
\arguments{
\item{irf_data}{The IRF data to be converted. Needs columns t, shock, and the response variables.}

\item{M_}{M_ object}
}
\value{
A data frame with columns t, shock, response variable name, response variable value, and units.
}
\description{
Convert IRF into a pretty format with names and units.
}
