% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_moo.R
\name{read_moo}
\alias{read_moo}
\title{Read/build a Moo object}
\usage{
read_moo(
  irf,
  model_name = NULL,
  meta = NULL,
  param_df = data.frame(),
  subset_vars = FALSE
)
}
\arguments{
\item{irf}{A long data frame with columns `t` (time), `resp_var` (response variable), `shock` (shock variable), and `value` (IRF value).}

\item{model_name}{Optional model name (used in pretty IRF outputs).}

\item{meta}{optional path to variable metadata (excel file with columns
dynare_name, display_name, units, unit_symbol, scale_formula [a formula for rescaling the IRF].
Additional metadata columns can be provided (for example, I add default [whether this variable is shown by default],
alt_paths [T/F whether to  include on alt paths module]).}

\item{param_df}{One-row data frame of model parameters used to
evaluate scaling formulas. Default is an empty dataframe.}

\item{subset_vars}{Boolean. if TRUE, clear the IRF dataframe of all variables not in subset_vars.}
}
\value{
A full MOO pair of class `custom_moo`, containing `M_` and `oo_`.
}
\description{
Convenience wrapper around [custom_moo()].
}
