% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dynare.R
\name{repack}
\alias{repack}
\title{Repack MATLAB-loaded objects with updated names.}
\usage{
repack(obj)
}
\arguments{
\item{obj}{Object loaded from MATLAB.}
}
\value{
Repacked object with restored names.
}
\description{
Repack MATLAB-loaded objects with updated names.
}
