% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_irf.R
\name{rescale}
\alias{rescale}
\title{Rescale an arbitrary matrix/dataframe (e.g. IRF) using metadata-defined scaling.}
\usage{
rescale(obj, M_)
}
\arguments{
\item{obj}{Object with columns matching `M_$varmeta$dynare_name`.}

\item{M_}{Dynare or custom model structure containing `varmeta` with scaling
definitions.}
}
\value{
Rescaled object with the same shape as input.
}
\description{
Rescale an arbitrary matrix/dataframe (e.g. IRF) using metadata-defined scaling.
}
