% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_irf.R
\name{rescale_irf}
\alias{rescale_irf}
\title{Rescale IRF using function or rescaling factor provided by user in metadata.}
\usage{
rescale_irf(irf, M_)
}
\arguments{
\item{irf}{IRF data frame or matrix with columns matching `M_$endo.names`.}

\item{M_}{Dynare or custom model structure containing `varmeta` with scaling
definitions.}
}
\value{
Rescaled IRF object with the same shape as input.
}
\description{
Rescale IRF using function or rescaling factor provided by user in metadata.
}
