% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_alt_paths_tab.R
\name{ServerTab_AltPaths}
\alias{ServerTab_AltPaths}
\title{Serve Alternative Policy Paths using configuration and public ezdyn APIs.}
\usage{
ServerTab_AltPaths(id, config)
}
\arguments{
\item{id}{Module namespace.}

\item{config}{Validated dashboard configuration.}
}
\value{
Invisibly, `NULL`.
}
\description{
Serve Alternative Policy Paths using configuration and public ezdyn APIs.
}
