% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_dictionary_tab.R
\name{ServerTab_Dictionary}
\alias{ServerTab_Dictionary}
\title{Serve the namespaced Variable Dictionary tab.}
\usage{
ServerTab_Dictionary(id, config)
}
\arguments{
\item{id}{Module namespace.}

\item{config}{Validated dashboard configuration.}
}
\value{
Invisibly, `NULL`.
}
\description{
Serve the namespaced Variable Dictionary tab.
}
