% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_optimal_policy_tab.R
\name{ServerTab_OptimalPolicy}
\alias{ServerTab_OptimalPolicy}
\title{Serve the Optimal Policy tab.}
\usage{
ServerTab_OptimalPolicy(id, config)
}
\arguments{
\item{id}{Module namespace.}

\item{config}{Validated dashboard configuration.}
}
\value{
Invisibly, `NULL`.
}
\description{
Serve the Optimal Policy tab.
}
