% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_overview_tab.R
\name{ServerTab_Overview}
\alias{ServerTab_Overview}
\title{Serve the namespaced Overview tab.}
\usage{
ServerTab_Overview(id, overview_config)
}
\arguments{
\item{id}{Module namespace.}

\item{overview_config}{Validated Overview configuration.}
}
\value{
Invisibly, `NULL`.
}
\description{
Serve the namespaced Overview tab.
}
