% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_alt_paths_tab.R
\name{UITab_AltPaths}
\alias{UITab_AltPaths}
\title{Build Alternative Policy Paths UI from configuration.}
\usage{
UITab_AltPaths(id, config)
}
\arguments{
\item{id}{Module namespace.}

\item{config}{Validated dashboard configuration.}
}
\value{
Shiny UI.
}
\description{
Build Alternative Policy Paths UI from configuration.
}
