% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_dictionary_tab.R
\name{UITab_Dictionary}
\alias{UITab_Dictionary}
\title{Build the namespaced Variable Dictionary tab UI.}
\usage{
UITab_Dictionary(id, config)
}
\arguments{
\item{id}{Module namespace.}

\item{config}{Validated dashboard configuration.}
}
\value{
Shiny UI.
}
\description{
Build the namespaced Variable Dictionary tab UI.
}
