% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_optimal_policy_tab.R
\name{UITab_OptimalPolicy}
\alias{UITab_OptimalPolicy}
\title{Build the Optimal Policy tab UI.}
\usage{
UITab_OptimalPolicy(id, config)
}
\arguments{
\item{id}{Module namespace.}

\item{config}{Validated dashboard configuration.}
}
\value{
Shiny UI.
}
\description{
Build the Optimal Policy tab UI.
}
