% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_overview_tab.R
\name{UITab_Overview}
\alias{UITab_Overview}
\title{Build the namespaced Overview tab UI.}
\usage{
UITab_Overview(id, overview_config)
}
\arguments{
\item{id}{Module namespace.}

\item{overview_config}{Validated Overview configuration.}
}
\value{
Shiny UI.
}
\description{
Build the namespaced Overview tab UI.
}
