% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dynare.R
\name{unpack_oo}
\alias{unpack_oo}
\title{Unpack oo_ object from .mat file.}
\usage{
unpack_oo(oo_)
}
\arguments{
\item{oo_}{Dynare `oo_` object loaded from MATLAB.}
}
\value{
Unpacked `oo_` list with restored names.
}
\description{
Unpack oo_ object from .mat file.
}
