% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plot.R
\name{plot_pretty_graph}
\alias{plot_pretty_graph}
\alias{irf_plot_pretty_graph}
\title{Build a graph from pretty IRF or alternative-path data.}
\usage{
plot_pretty_graph(
  df,
  line_style = c("combined", "split"),
  plotter = c("ggrba", "ggplot"),
  custom_title = NULL,
  baseline_name = "Baseline",
  baseline_colour = "royalblue"
)

irf_plot_pretty_graph(...)
}
\arguments{
\item{df}{Data frame created by [pretty_irf()] or [get_alt_paths()].}

\item{line_style}{How to treat shock/model combinations for lines.}

\item{plotter}{Plotting backend. `"ggrba"` (default) or `"ggplot"`.}

\item{custom_title}{Provide custom title if TRUE.}

\item{baseline_name}{Label used for an alternative-path baseline.}

\item{baseline_colour}{Colour used for an alternative-path baseline.}
}
\value{
List containing the `graph` and `graph_fname`.
}
\description{
Build a graph from pretty IRF or alternative-path data.
}
